+++ config/filament-shield.php
@@ -247,7 +247,6 @@ return [
     'custom_permissions' => [
         'Approve:CmsPage',
         'Approve:CmsPost',
-        'Manage:CodeInjection',
         'SubmitForReview:CmsPage',
         'SubmitForReview:CmsPost',
         'ViewRevisions:CmsPage',