+++ packages/tallcms/cms/resources/views/filament/pages/code-injection.blade.php
@@ -1,9 +1,0 @@
-<x-filament-panels::page>
-    <form wire:submit="save">
-        {{ $this->form }}
-
-        <x-filament::button type="submit" class="mt-6">
-            Save
-        </x-filament::button>
-    </form>
-</x-filament-panels::page>