+++ packages/tallcms/cms/src/TallCmsServiceProvider.php
@@ -244,7 +244,6 @@ class TallCmsServiceProvider extends PackageServiceProvider
         'App\\Policies\\UserPolicy' => Policies\UserPolicy::class,
 
         // Filament Pages
-        'App\\Filament\\Pages\\CodeInjection' => Filament\Pages\CodeInjection::class,
         'App\\Filament\\Pages\\MenuItemsManager' => Filament\Pages\MenuItemsManager::class,
         'App\\Filament\\Pages\\PluginLicenses' => Filament\Pages\PluginLicenses::class,
         'App\\Filament\\Pages\\PluginManager' => Filament\Pages\PluginManager::class,
@@ -337,7 +336,6 @@ class TallCmsServiceProvider extends PackageServiceProvider
         $cmsPermissions = [
             'Approve:CmsPage',
             'Approve:CmsPost',
-            'Manage:CodeInjection',
             'SubmitForReview:CmsPage',
             'SubmitForReview:CmsPost',
             'ViewRevisions:CmsPage',